src/main/webapp/app/entities/fonc-agent/delete/fonc-agent-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { FoncAgentService } from '../service/fonc-agent.service';

import { FoncAgentDeleteDialogComponent } from './fonc-agent-delete-dialog.component';

describe('FoncAgent Management Delete Component', () => {
  let comp: FoncAgentDeleteDialogComponent;
  let fixture: ComponentFixture<FoncAgentDeleteDialogComponent>;
  let service: FoncAgentService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [FoncAgentDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(FoncAgentDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(FoncAgentDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(FoncAgentService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
